editor/libeditor/tests/test_bug417418.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=417418
-->
<head>
  <title>Test for Bug 417418</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=417418">Mozilla Bug 417418</a>
<div id="display" contenteditable="true">
<p id="coin">first paragraph</p>
<p>second paragraph. <img id="img" src="green.png"></p>
</div>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 417418 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(runTest);

function resetSelection() {
  window.getSelection().collapse(document.getElementById("coin"), 0);
}

function runTest() {
    var rightClick = {type: 'mousedown', button: 2},
        singleClick = {type: 'mousedown', button: 0};
    var selection = window.getSelection();

    var div = document.getElementById('display');
    var img = document.getElementById('img');
    var divRect = div.getBoundingClientRect();
    var imgselected;

    resetSelection();
    synthesizeMouse(div, divRect.width - 1, divRect.height - 1, rightClick);
    ok(selection.isCollapsed, "selection is not collapsed");

    resetSelection();
    synthesizeMouse(div, divRect.width - 1, divRect.height - 1, singleClick);
    ok(selection.isCollapsed, "selection is not collapsed");

    resetSelection();
    synthesizeMouseAtCenter(img, rightClick);
    imgselected = selection.anchorNode == img.parentNode &&
                      selection.anchorOffset === 1 &&
                      selection.rangeCount === 1;
    ok(imgselected, "image is not selected");

    resetSelection();
    synthesizeMouseAtCenter(img, singleClick);
    imgselected = selection.anchorNode == img.parentNode &&
                      selection.anchorOffset === 1 &&
                      selection.rangeCount === 1;
    ok(imgselected, "image is not selected");

    SimpleTest.finish();
}


</script>
</pre>
</body>
</html>
